+++ serde_amqp/benches/serialize.rs
@@ -1,10 +1,8 @@
 #![allow(clippy::all)]
 
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use rand::{
-    distributions::{Alphanumeric, DistString},
-    Rng, RngCore,
-};
+use rand::{distr::SampleString, Rng, RngCore};
+use rand_distr::Alphanumeric;
 use serde_amqp::primitives::{Binary, Dec128, Dec32, Dec64, Timestamp};
 
 fn criterion_benchmark(c: &mut Criterion) {
@@ -100,78 +98,78 @@ fn criterion_benchmark(c: &mut Criterion) {
     });
 
     let mut value = vec![0u8; 16];
-    rand::thread_rng().fill_bytes(&mut value);
+    rand::rng().fill_bytes(&mut value);
     let value = Binary::from(value);
     c.bench_function("serialize Binary 16B", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
     let mut value = vec![0u8; 64];
-    rand::thread_rng().fill_bytes(&mut value);
+    rand::rng().fill_bytes(&mut value);
     let value = Binary::from(value);
     c.bench_function("serialize Binary 64B", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
     let mut value = vec![0u8; 256];
-    rand::thread_rng().fill_bytes(&mut value);
+    rand::rng().fill_bytes(&mut value);
     let value = Binary::from(value);
     c.bench_function("serialize Binary 256B", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
     let mut value = vec![0u8; 1024];
-    rand::thread_rng().fill_bytes(&mut value);
+    rand::rng().fill_bytes(&mut value);
     let value = Binary::from(value);
     c.bench_function("serialize Binary 1kB", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
     let mut value = vec![0u8; 1024 * 1024];
-    rand::thread_rng().fill_bytes(&mut value);
+    rand::rng().fill_bytes(&mut value);
     let value = Binary::from(value);
     c.bench_function("serialize Binary 1MB", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
     let mut value = vec![0u8; 10 * 1024 * 1024];
-    rand::thread_rng().fill_bytes(&mut value);
+    rand::rng().fill_bytes(&mut value);
     let value = Binary::from(value);
     c.bench_function("serialize Binary 10MB", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
-    let value = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
+    let value = Alphanumeric.sample_string(&mut rand::rng(), 16);
     let value = String::from(value);
     c.bench_function("serialize String 16B", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
-    let value = Alphanumeric.sample_string(&mut rand::thread_rng(), 64);
+    let value = Alphanumeric.sample_string(&mut rand::rng(), 64);
     let value = String::from(value);
     c.bench_function("serialize String 64B", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
-    let value = Alphanumeric.sample_string(&mut rand::thread_rng(), 256);
+    let value = Alphanumeric.sample_string(&mut rand::rng(), 256);
     let value = String::from(value);
     c.bench_function("serialize String 256B", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
-    let value = Alphanumeric.sample_string(&mut rand::thread_rng(), 1024);
+    let value = Alphanumeric.sample_string(&mut rand::rng(), 1024);
     let value = String::from(value);
     c.bench_function("serialize String 1kB", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
-    let value = Alphanumeric.sample_string(&mut rand::thread_rng(), 1024 * 1024);
+    let value = Alphanumeric.sample_string(&mut rand::rng(), 1024 * 1024);
     let value = String::from(value);
     c.bench_function("serialize String 1MB", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
-    let value = Alphanumeric.sample_string(&mut rand::thread_rng(), 10 * 1024 * 1024);
+    let value = Alphanumeric.sample_string(&mut rand::rng(), 10 * 1024 * 1024);
     let value = String::from(value);
     c.bench_function("serialize String 10MB", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
@@ -179,42 +177,42 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     // 16 bytes of u64
     let mut value = vec![0u64; 16 / std::mem::size_of::<u64>()];
-    rand::thread_rng().fill(&mut value[..]);
+    rand::rng().fill(&mut value[..]);
     c.bench_function("serialize List<u64> 16B", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
     // 64 bytes of u64
     let mut value = vec![0u64; 64 / std::mem::size_of::<u64>()];
-    rand::thread_rng().fill(&mut value[..]);
+    rand::rng().fill(&mut value[..]);
     c.bench_function("serialize List<u64> 64B", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
     // 256 bytes of u64
     let mut value = vec![0u64; 256 / std::mem::size_of::<u64>()];
-    rand::thread_rng().fill(&mut value[..]);
+    rand::rng().fill(&mut value[..]);
     c.bench_function("serialize List<u64> 256B", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
     // 1kB of u64
     let mut value = vec![0u64; 1024 / std::mem::size_of::<u64>()];
-    rand::thread_rng().fill(&mut value[..]);
+    rand::rng().fill(&mut value[..]);
     c.bench_function("serialize List<u64> 1kB", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
     // 1MB of u64
     let mut value = vec![0u64; 1024 * 1024 / std::mem::size_of::<u64>()];
-    rand::thread_rng().fill(&mut value[..]);
+    rand::rng().fill(&mut value[..]);
     c.bench_function("serialize List<u64> 1MB", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
 
     // 10MB of u64
     let mut value = vec![0u64; 10 * 1024 * 1024 / std::mem::size_of::<u64>()];
-    rand::thread_rng().fill(&mut value[..]);
+    rand::rng().fill(&mut value[..]);
     c.bench_function("serialize List<u64> 10MB", |b| {
         b.iter(|| serde_amqp::to_vec(black_box(&value)).unwrap())
     });
@@ -222,7 +220,7 @@ fn criterion_benchmark(c: &mut Criterion) {
     // list of strings
     // 10 strings of 16B
     let value = (0..10)
-        .map(|_| Alphanumeric.sample_string(&mut rand::thread_rng(), 16))
+        .map(|_| Alphanumeric.sample_string(&mut rand::rng(), 16))
         .map(String::from)
         .collect::<Vec<_>>();
     c.bench_function("serialize List<String> 10x16B", |b| {
@@ -231,7 +229,7 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     // 100 strings of 16B
     let value = (0..100)
-        .map(|_| Alphanumeric.sample_string(&mut rand::thread_rng(), 16))
+        .map(|_| Alphanumeric.sample_string(&mut rand::rng(), 16))
         .map(String::from)
         .collect::<Vec<_>>();
     c.bench_function("serialize List<String> 100x16B", |b| {
@@ -240,7 +238,7 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     // 1000 strings of 16B
     let value = (0..1000)
-        .map(|_| Alphanumeric.sample_string(&mut rand::thread_rng(), 16))
+        .map(|_| Alphanumeric.sample_string(&mut rand::rng(), 16))
         .map(String::from)
         .collect::<Vec<_>>();
     c.bench_function("serialize List<String> 1000x16B", |b| {
@@ -249,7 +247,7 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     // 10 strings of 1kB
     let value = (0..10)
-        .map(|_| Alphanumeric.sample_string(&mut rand::thread_rng(), 1024))
+        .map(|_| Alphanumeric.sample_string(&mut rand::rng(), 1024))
         .map(String::from)
         .collect::<Vec<_>>();
     c.bench_function("serialize List<String> 10x1kB", |b| {
@@ -258,7 +256,7 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     // 100 strings of 1kB
     let value = (0..100)
-        .map(|_| Alphanumeric.sample_string(&mut rand::thread_rng(), 1024))
+        .map(|_| Alphanumeric.sample_string(&mut rand::rng(), 1024))
         .map(String::from)
         .collect::<Vec<_>>();
     c.bench_function("serialize List<String> 100x1kB", |b| {
@@ -267,7 +265,7 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     // 1000 strings of 1kB
     let value = (0..1000)
-        .map(|_| Alphanumeric.sample_string(&mut rand::thread_rng(), 1024))
+        .map(|_| Alphanumeric.sample_string(&mut rand::rng(), 1024))
         .map(String::from)
         .collect::<Vec<_>>();
     c.bench_function("serialize List<String> 1000x1kB", |b| {
@@ -276,7 +274,7 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     // 10 strings of 1MB
     let value = (0..10)
-        .map(|_| Alphanumeric.sample_string(&mut rand::thread_rng(), 1024 * 1024))
+        .map(|_| Alphanumeric.sample_string(&mut rand::rng(), 1024 * 1024))
         .map(String::from)
         .collect::<Vec<_>>();
     c.bench_function("serialize List<String> 10x1MB", |b| {
@@ -322,9 +320,9 @@ fn criterion_benchmark(c: &mut Criterion) {
     // Map of 10 16B String -> 16B String
     let value = (0..10)
         .map(|_| {
-            let key = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
+            let key = Alphanumeric.sample_string(&mut rand::rng(), 16);
             let key = String::from(key);
-            let value = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
+            let value = Alphanumeric.sample_string(&mut rand::rng(), 16);
             let value = String::from(value);
             (key, value)
         })
@@ -336,9 +334,9 @@ fn criterion_benchmark(c: &mut Criterion) {
     // Map of 100 16B String -> 16B String
     let value = (0..100)
         .map(|_| {
-            let key = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
+            let key = Alphanumeric.sample_string(&mut rand::rng(), 16);
             let key = String::from(key);
-            let value = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
+            let value = Alphanumeric.sample_string(&mut rand::rng(), 16);
             let value = String::from(value);
             (key, value)
         })
@@ -350,9 +348,9 @@ fn criterion_benchmark(c: &mut Criterion) {
     // Map of 1000 16B String -> 16B String
     let value = (0..1000)
         .map(|_| {
-            let key = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
+            let key = Alphanumeric.sample_string(&mut rand::rng(), 16);
             let key = String::from(key);
-            let value = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
+            let value = Alphanumeric.sample_string(&mut rand::rng(), 16);
             let value = String::from(value);
             (key, value)
         })